models/sampling/attention_based_sampler.py
"""
Attention-Based Sampling implementation for ProteinFlex.
Implements structure-aware attention for protein generation.
"""

import torch
import torch.nn as nn
import torch.nn.functional as F
import math
from typing import Dict, Tuple, Optional

class AttentionBasedSampler(nn.Module):
    def __init__(
        self,
        feature_dim: int = 768,
        hidden_dim: int = 512,
        num_heads: int = 8,
        num_layers: int = 6,
        dropout: float = 0.1
    ):
        """
        Initialize Attention-Based Sampler.

        Args:
            feature_dim: Dimension of protein features
            hidden_dim: Hidden dimension
            num_heads: Number of attention heads
            num_layers: Number of transformer layers
            dropout: Dropout rate
        """
        super().__init__()
        self.feature_dim = feature_dim
        self.hidden_dim = hidden_dim
        self.num_heads = num_heads
        self.head_dim = hidden_dim // num_heads

        # Input projection
        self.input_proj = nn.Linear(feature_dim, hidden_dim)

        # Transformer layers
        self.layers = nn.ModuleList([
            TransformerLayer(hidden_dim, num_heads, dropout)
            for _ in range(num_layers)
        ])

        # Output projection
        self.output_proj = nn.Linear(hidden_dim, feature_dim)

    def forward(
        self,
        x: torch.Tensor,
        structure_bias: Optional[torch.Tensor] = None
    ) -> torch.Tensor:
        """
        Forward pass.

        Args:
            x: Input features [batch_size, seq_len, feature_dim]
            structure_bias: Optional structure information [batch_size, seq_len, seq_len]

        Returns:
            Updated features [batch_size, seq_len, feature_dim]
        """
        # Project input
        h = self.input_proj(x)

        # Apply transformer layers
        for layer in self.layers:
            h = layer(h, structure_bias)

        # Project output
        return self.output_proj(h)

    def sample(
        self,
        batch_size: int,
        seq_len: int,
        device: torch.device,
        temperature: float = 1.0,
        structure_bias: Optional[torch.Tensor] = None
    ) -> torch.Tensor:
        """
        Generate protein features using attention-based sampling.

        Args:
            batch_size: Number of samples to generate
            seq_len: Sequence length
            device: Device to generate on
            temperature: Sampling temperature
            structure_bias: Optional structure guidance [batch_size, seq_len, seq_len]

        Returns:
            Generated features [batch_size, seq_len, feature_dim]
        """
        # Initialize random features
        x = torch.randn(
            batch_size, seq_len, self.feature_dim,
            device=device
        ) * temperature

        # Refine through attention
        return self.forward(x, structure_bias)

    def compute_loss(
        self,
        pred_features: torch.Tensor,
        target_features: torch.Tensor,
        structure_bias: Optional[torch.Tensor] = None
    ) -> torch.Tensor:
        """
        Compute training loss.

        Args:
            pred_features: Predicted features [batch_size, seq_len, feature_dim]
            target_features: Target features [batch_size, seq_len, feature_dim]
            structure_bias: Optional structure information [batch_size, seq_len, seq_len]

        Returns:
            Loss value
        """
        # Feature reconstruction loss
        feature_loss = F.mse_loss(pred_features, target_features)

        # Structure-aware loss if bias provided
        if structure_bias is not None:
            pred_dist = torch.cdist(pred_features, pred_features)
            target_dist = torch.cdist(target_features, target_features)
            structure_loss = F.mse_loss(pred_dist, target_dist)
            return feature_loss + structure_loss

        return feature_loss

class TransformerLayer(nn.Module):
    def __init__(
        self,
        hidden_dim: int,
        num_heads: int,
        dropout: float = 0.1
    ):
        """Initialize transformer layer."""
        super().__init__()
        self.attention = MultiHeadAttention(hidden_dim, num_heads, dropout)
        self.norm1 = nn.LayerNorm(hidden_dim)
        self.norm2 = nn.LayerNorm(hidden_dim)

        self.ff = nn.Sequential(
            nn.Linear(hidden_dim, hidden_dim * 4),
            nn.ReLU(),
            nn.Dropout(dropout),
            nn.Linear(hidden_dim * 4, hidden_dim)
        )

        self.dropout = nn.Dropout(dropout)

    def forward(
        self,
        x: torch.Tensor,
        structure_bias: Optional[torch.Tensor] = None
    ) -> torch.Tensor:
        """
        Forward pass for transformer layer.


        Args:
            x: Input features [batch_size, seq_len, hidden_dim]
            structure_bias: Optional structure information [batch_size, seq_len, seq_len]
        """
        # Self-attention
        attended = self.attention(x, x, x, structure_bias)
        x = self.norm1(x + self.dropout(attended))

        # Feed-forward
        ff_out = self.ff(x)
        return self.norm2(x + self.dropout(ff_out))

class MultiHeadAttention(nn.Module):
    def __init__(
        self,
        hidden_dim: int,
        num_heads: int,
        dropout: float = 0.1
    ):
        """Initialize multi-head attention."""
        super().__init__()
        self.hidden_dim = hidden_dim
        self.num_heads = num_heads
        self.head_dim = hidden_dim // num_heads

        self.q_proj = nn.Linear(hidden_dim, hidden_dim)
        self.k_proj = nn.Linear(hidden_dim, hidden_dim)
        self.v_proj = nn.Linear(hidden_dim, hidden_dim)
        self.out_proj = nn.Linear(hidden_dim, hidden_dim)

        self.dropout = nn.Dropout(dropout)

    def forward(
        self,
        query: torch.Tensor,
        key: torch.Tensor,
        value: torch.Tensor,
        structure_bias: Optional[torch.Tensor] = None
    ) -> torch.Tensor:
        """
        Multi-head attention forward pass.

        Args:
            query: Query tensor [batch_size, seq_len, hidden_dim]
            key: Key tensor [batch_size, seq_len, hidden_dim]
            value: Value tensor [batch_size, seq_len, hidden_dim]
            structure_bias: Optional structure information [batch_size, seq_len, seq_len]
        """
        batch_size, seq_len = query.shape[:2]

        # Project and reshape for attention heads
        q = self.q_proj(query).view(batch_size, seq_len, self.num_heads, self.head_dim).transpose(1, 2)
        k = self.k_proj(key).view(batch_size, seq_len, self.num_heads, self.head_dim).transpose(1, 2)
        v = self.v_proj(value).view(batch_size, seq_len, self.num_heads, self.head_dim).transpose(1, 2)

        # Compute attention scores
        scores = torch.matmul(q, k.transpose(-2, -1)) / math.sqrt(self.head_dim)

        # Add structure bias if provided
        if structure_bias is not None:
            scores = scores + structure_bias.unsqueeze(1)

        # Apply attention
        attn = F.softmax(scores, dim=-1)
        attn = self.dropout(attn)

        # Get output
        out = torch.matmul(attn, v)
        out = out.transpose(1, 2).contiguous().view(batch_size, seq_len, self.hidden_dim)

        return self.out_proj(out)
